requirements.txt
# Adaptive AI Firewall – dependencies
# Install: pip install -r requirements.txt
# Then:    python -m spacy download en_core_web_lg

sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.0
openai>=1.30.0
python-dotenv>=1.0.0
numpy>=1.26.0
